Changes requeted by Jean Carlo
