--- Semana 10 Ejercicios/functions/allstudents.py
@@ -1,7 +1,0 @@
-from .students import amount_of_students
-
-
-def total_entered_students():
-    print(amount_of_students)
-
-
