+++ Semana 10 Ejercicios/functions/students.py
@@ -1,6 +1,5 @@
-amount_of_students = []
-
 def students():
+    amount_of_students = []
     while True:
         name = str(input("Ingrese el nombre del estudiante: "))
         section = input("Ingrese la sección (Ejemplo 11B): ")
@@ -69,3 +68,50 @@ def students():
         else: 
             yes_or_menu == 2
             break
+
+    return amount_of_students
+
+
+def total_entered_students(students_list):
+    for student in students_list:
+        print(student)
+
+
+def average_grades(students_list):
+    averages = []
+    for student in students_list:
+        grades = [
+            student['Nota de Español'],
+            student['Nota de Inglés'],
+            student['Nota de Sociales'],
+            student['Nota de Ciencias']
+        ]
+
+        average_student_grades = sum(grades) / len(grades)
+        averages.append({
+            "Nombre" : student['Nombre'],
+            "Promedio" : average_student_grades
+        })
+    return averages
+
+def top_3_average(student_list):
+    if not student_list:         
+        return []
+    promedios = average_grades(student_list)
+    promedios_ordenados = sorted(promedios, key=lambda x: x["Promedio"], reverse=True)
+    return promedios_ordenados[:3]
+
+def sum_all_averages(student_list):
+    sum_of_averages = []
+    for student in student_list:
+        grades = [
+            student['Nota de Español'],
+            student['Nota de Inglés'],
+            student['Nota de Sociales'],
+            student['Nota de Ciencias']
+        ]
+
+        average_student_grades = sum(grades) / len(grades)
+        sum_of_averages.append(average_student_grades)
+        sum_averages = sum(sum_of_averages) / len(sum_of_averages)
+    return sum_averages
