+++ students.csv
@@ -1,5 +1,5 @@
 Nombre,Sección,Nota de Español,Nota de Inglés,Nota de Sociales,Nota de Ciencias
-Greg,11A,100,89,70,88
-Adri,11A,100,90,89,78
-Dylan,11A,88,98,75,77
-Jane,11A,100,68,70,88
+Greg,11A,89,70,98,100
+Adri,11A,89,98,100,100
+Dylan,98,78,88,100,98
+Jane,60,70,85,95,60
